BackendAPI/TrimAPI/requirements.txt
google-cloud-storage
ffmpeg-python